= empty comments, bumps version for private branch
we added a gpg key user and want to make a new tag out of it on private
branch, but keep the tag version numbers the same, so we did a change
there, but not one here.  fix that by adding a comment line so we can
make a new reltag.  then rebase private branch off it, and then reltag
match on the private branch