- user: force email to be defined in local config
the email of a commit needs to change depending on the chosen identity:
for work, the work email should be used.  otherwise things get pushed
that have the wrong identity.  use the git-2.8.0 new useConfigOnly to
require email to be set, and then unset it in our global user config so
that git will fail unless it's been locally defined.  this requires all
repositories to explicitly define which user to commit as