+ alias: lsa: add tree listing capability (ala lsa)
the argument parsing isn't very nice, because both branch and files to
list need to be provided, but which is first? one arg we probably want
to default to '@' but maybe we want to default to '/' instead which
doesn't actually work.  and we throw in recursive... so we really need a
shell wrapper that does getopt-like parsing.  for now this will have to
do, at least it gets us up and using the command.