+ alias: noff: use to force a merge commit
even if you would otherwise fast forward.  this lets rebase-only
appending branches still have a merge commit which can group the
commits.  otherwise the only possibility when using such a strategy is
one continuous chain of commits, but sometimes we want to keep the
information about groups of commits which might all add a feature (ie
when pulling a rebased feature branch)