- alias: move verbose commits into config instead of cli option
we always like commit message edits to have the patch shown in commts,
to avoid having to go get the patch to write the commit message.
this is under control of verbose flag.  we can give this on command line
(which we were doing via 'ci' commit alias), or in config so it's
permanent.  can't think of a reason why we would ever want this off so
move to config