+ alias: lastn: lasttt with optional count arg, lasttt is now 30
we very commonly do "git lasttt -30", so much so that, instead of using
"git lasttt" to mean "display arbitrarily many, and allow another
argument such as the count", we'll instead make it mean "display 30",
and change the "arbitrarily many" to a "lastn" which takes a number
argument and display that many.  if not provided, no argument will be
given, meaning it's the same as lasttt used to be (and has the advantage
of being one less character in the no-count case, and two less in the
count-provided case ("lasttt -100" -> "lastn 100").