/ alias: lastn: fix incorrect recognition of first arg being a count
it would fail on commands like "git lastn 702f7923f..79952b9c6" because
this would be seen as a number due to "${1##[0-9]*}" just meaning
"starts with digit" not "whole word is digit"

to match the whole word as digits correctly, we have to use bash regex
operator.  there might be a way in original bourne, but could not come
up with it off the top of my head.  callout to bash was easy enough, as
we already had examples in the same file with the right syntax