+ alias: lsa: take arbitrary args, and default to HEAD if not given
we test the given $1 to see if it resolves as a revision ("cat-file -e"
is intended for this purpose, ie to test whether a "git object" exists),
and if it does, use it; otherwise default to the HEAD object (ie "@").

the next arg if not supplied defaults to '.' because that's the usual
expectation of our 'lsa' wrapper (ie "ls") used on files.

finally all the remaining args are given, rather than just $2 as before