/ alias: fix all shellout aliases to handle zero or many args correctly
there is a quirk with command line invocation of the shell using '-c'
and how git creates the arguments, which can be seen here:

     $ bash -c 'f() { echo n:$# a:$@; }; f'
    n:0 a:

     $ bash -c 'f() { echo n:$# a:$@; }; f $@' 1
    n:0 a:

     $ bash -c 'f() { echo n:$# a:$@; }; f $@' 1 2
    n:1 a:2

     $ bash -c 'f() { echo n:$# a:$@; }; f $@' 1 2 3
    n:2 a:2 3

notice how (1) when no args supplied, it's the same as 1 arg supplied,
and (2) there is one less arg in $# than expected.  this is because of
the way '-c' works, as shown in the bash man page:

      bash [options] [command_string | file]

    If the -c option is present, then commands are read from  the first
    non-option argument command_string.  If there are arguments after
    the command_string, the first argument is assigned to $0 and any
    remaining arguments are assigned to the positional parameters.  The
    assignment to $0 sets the name of the shell

it's the same for dash:

       dash -c command_string [command_name [argument ...]]

    Read commands from the command_string operand instead of from the
    standard input.  Special parameter 0 will be set from the
    command_name operand and the positional parameters ($1, $2, etc.)
    set from the remaining argument operands

so the shell takes the first arg as the string of commands, the second
arg as $0, and the rest of the args as $1 - $N

Now git is actually calling out to /bin/sh first and then to bash when
we use '!bash', and when there are no args, it uses:

    sh -c "foo bar" "foo bar"

and when there are args, in git:run-command.c:prepare_shell_command():

    strvec_push(out, "sh");
    strvec_push(out, "-c");

    /*
     * If we have no extra arguments, we do not even need to
     * bother with the "$@" magic.
     */
    if (!argv[1])
            strvec_push(out, argv[0]);
    else
            strvec_pushf(out, "%s \"$@\"", argv[0]);

    strvec_pushv(out, argv);

so when args are present:

    sh -c "foo bar \"$@\"" "foo bar" arg1 arg2

so if we had an alias like so:

    testalias = "!bash -c 'echo \"0:$0,1:$1,2:$2\"'"

then without args:

     $ strace -qq -s 999 -fe execve git testalias 2>&1 | grep execve | fmt
    execve("/usr/bin/git", ["git", "testalias"], 0x7ffe4a7f1de0 /* 80 vars
       */) = 0
    [pid 523924] execve("/bin/sh", ["/bin/sh", "-c", "bash -c 'echo
       \"0:/bin/bash,1:,2:\"'", "bash -c 'echo \"0:/bin/bash,1:,2:\"'"],
       0x5639c183d1d0 /* 82 vars */) = 0
    [pid 523925] execve("/bin/bash", ["bash", "-c", "echo
       \"0:/bin/bash,1:,2:\""], 0x55678def3378 /* 82 vars */) = 0
    0:bash,1:,2:

but with args:

     $ strace -qq -s 999 -fe execve git testalias 1 2 2>&1 | grep execve | fmt
    execve("/usr/bin/git", ["git", "testalias", "1", "2"], 0x7ffe92160ce0 /*
       80 vars */) = 0
    [pid 523836] execve("/bin/sh", ["/bin/sh", "-c", "bash -c 'echo
       \"0:/bin/bash,1:,2:\"' \"\"", "bash -c 'echo \"0:/bin/bash,1:,2:\"'",
       "1", "2"], 0x5649603f6210 /* <82 vars */) = 0
    [pid 523837] execve("/bin/bash", ["bash", "-c", "echo
       \"0:/bin/bash,1:,2:\"", "1", "2"], 0x55f52f19a378 /* 82 vars */) = 0
    0:1,1:2,2:

the solution to make it work no matter how many args (including none) is
to add a dummy arg after the fully quoted command, ie

    bash -c 'cmd' dummy

which will then always cause the "$@" form to be used, and we can
control that $1..N will always be the args given as parameters to the
command, as expected.

for the dummy, we simply use the alias name prefaced by 'git-' as though
it were in fact an external script implementation.