^ rc: pager: rely on $LESS for flags, but ensure that '-X' is set
we have just done a bunch of commits to .bash/env to have more sane
defaults in 'less', in particular we are using '-F' now so that '-E' is
not necessary, and in fact it's not desired for long-output cases,
because we would go to the bottom and it would promptly exit, which was
annoying.  with '-F' we get no paging when less than a screenful, but
keep having to explicitly exit when more than a screenful, which is
usually what we want.

BUT we do retain forcing '-X' because when looking at 'git' output we
almost never would want to use secondary screen, because we're typically
looking at a diff or log files and want to keep that output shown on
control-z and after exit.  We don't put this in $LESS because in regular
terminal use, typically we DO want to use secondary scren, so that
control-z brings us back to whatever we were looking at before paging.