+ secrets: cf: smemsh: rewrite all github remotes as git@pseudo-host
all github users are git@github.com, and use their key to identify them.
but we use IdentityFile in ssh config, which uses Host match.
furthermore, we do prevent identity leak by using IdentitiesOnly so we
cannot just load and offer all our keys.  we have to provide the key in
IdentifyFile.  so to match different github accounts we provide some
pseudo-Hosts that differ in Host (which user speecifies) but not
HostName (github.com); this lets us select the [exclusive] key to offer
based on something we can put in the repo url.

this means we need to modify the 'origin' server url to use that domain,
rather than just "github.com" we need to use for Host portion of ssh
connect specifier, ie "github-personal", "github-work", etc

this patch does that automatically.  usage would be eg "gh fork
--remote" which sets 'upstream' remote to github https, we typically
only pull from there and don't rewrite that.  then we push to our own
'origin' which needs its url rewritten to the pseudo-Host.

  command: git cf-smemsh
  before: user, email
  after: user, email, origin

in order to make using forks with the paradigm easier