+ rc: fetch: parallel: set to 8 parallel fetches, speeds submodules
submodules can use the separate submodule.fetchJobs but setting
fetch.parallel configures it for both submodules and fetch --multi

this speeds up some repos that have a lot of submodules and allows
fetching in parallel.  we use 8 which is double the 4 that browsers
started using long ago, this was the Netscape limit for parallel
downloads, back when HTTP/1.0 used separate TCP sockets for each
connection.  8 seems reasonable to start with and we will titrate if
there is a problem found with this many in the field