/ aliases: shellouts: fix lsa, lastn unaware of chdir to gitroot
when invoking an aliased shell wrapper, git apparently changes directory
to the git root, as noted in git-config(1):

  Note that shell commands will be executed from the top-level directory
  of a repository, which may not necessarily be the current directory.
  GIT_PREFIX is set as returned by running git rev-parse --show-prefix
  from the original current directory. See git-rev-parse(1).

we discovered this when doing a "git lastn -- ." in a subdir of the
repo, and it would print all commits, rather than just the subset in
of commits from paths in that dir.  whereas manually spelling out "git
log --oneline -- ." was returning the expected results.  quite puzzling;
we did not consider that it would do a chdir, and the execve() args
looked exactly the same upon trace, so this was highly confusing at
first.  found it by doing an "strace --no-abbrev" and comparing the
full environment of the execve() between the two line by line in vim
diff mode.  PWD was different, leading to the "aha" moment.

this also explains a longstanding puzzlement as to why "git lsa" was
always working relative to the root, but we had just thought this must
have been intended behavior.  thankfully that problem is now fixed as
well.

solution is to chdir into $GIT_PREFIX as described in the aforementioned
manual...