= aliases: shellouts: do not indent long defs to avoid whitespace in ps
the spaces are not trimmed so they appear in 'ps' which is ugly.  we
left a couple infrequently-used aliases still in the file, these are
used with some frequency, especially lastn which is one of the most used
aliases and frequently have shells in the process table that had called
it.

probably these should just be installed as shell scripts rather than
inline, but maybe we'll do that later...